# RachanaSaneSwagger-farmWaterOrder-1.0.1-swagger.yaml
swagger: '2.0'
info:
  description: |
    This is a documentation for WaterOrder services
  version: "1.0.1"
  title: Swagger WaterOrder
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
# host: petstore.swagger.io
# basePath: /water
# schemes:
# - http
paths:
  /order:
    post:
      tags:
      - pet
      summary: Add a new waterOrder to the store
      operationId: addPet
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: waterOrder object that needs to be created for the farmer
        required: true
        schema:
          $ref: '#/definitions/WaterOrder'
      responses:
        200: 
          description: Successfully complete
          schema:
            $ref: '#/definitions/WaterOrder'
        400: 
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
    
  
  /order/{orderId}:
    get:
      summary: Find pet by orderId
      description: Returns a single water order
      produces:
      - application/json
      
      parameters:
      - name: orderId
        in: path
        description: ID of waterOrder to return
        required: true
        type: integer
        format: int64
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/WaterOrder'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
        404:
          description: waterid not found 
          schema:
            $ref: '#/definitions/Error'
 
    patch:
      
      summary: Cancel waterOrder by orderId
      description: Returns a single water order that is cancelled
      produces:
      - application/json
      
      parameters:
      - name: orderId
        in: path
        description: ID of waterOrder to cancel
        required: true
        type: integer
        format: int64
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/WaterOrder'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
        404:
          description: waterid not found 
          schema:
            $ref: '#/definitions/Error'
          
  /orders:
    get:
      
      summary: Get all water orders
      produces:
      - application/json
      responses:
        200:
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/WaterOrder'
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
  

definitions:
  WaterOrder:
    type: object
    required:
    - farmName
    - startDateTime
    - duration
    properties:
      id:
        type: integer
        format: int64
      startDateTime:
        type: string
        format: date-time
        example: doggie
      duration:
        type: string
        example: doggie  
      orderStatusDescription:
        type: string
        enum:
          - Order has been placed but not yet delivered
          - Order is being delivered right now
          - Order has been delivered
          - Order was cancelled before delivery
          
  Error:
    type: object
    
    properties:
      status:
        type: integer
        format: int32
      errors:
        type: array
        items:
          type: string
               
      
  
  

host: virtserver.swaggerhub.com
basePath: /RachanaSaneSwagger/farmWaterOrder/1.0.0
schemes:
 - https
 - http